Rename lib/translationunit.js to lib/translation-unit.js
